% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_box_scores.R
\name{load_box_scores}
\alias{load_box_scores}
\title{Load team-level box scores data.}
\usage{
load_box_scores()
}
\value{
A data set containing all available box scores at the team level, from games played since the 2014/15 season onwards.
}
\description{
Load team-level box scores data.
}
\examples{
box_scores_data <- load_box_scores()
}
