% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_box_scores_detailed.R
\name{load_box_scores_detailed}
\alias{load_box_scores_detailed}
\title{Load player-level box scores data.}
\usage{
load_box_scores_detailed()
}
\value{
A data set containing all available box scores at the player level, from games played since the 2014/15 season onwards.
}
\description{
Load player-level box scores data.
}
\examples{
box_scores_detailed_data <- load_box_scores_detailed()
}
