% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbp.R
\name{load_pbp}
\alias{load_pbp}
\title{Load play-by-play data.}
\usage{
load_pbp()
}
\value{
A data set containing all available play-by-play data, from games played since the 2014/15 season onwards.
}
\description{
Load play-by-play data.
}
\examples{
pbp_data <- load_pbp()
}
