% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shots.R
\name{load_shots}
\alias{load_shots}
\title{Load shots data.}
\usage{
load_shots()
}
\value{
A data set containing all available shots data, from games played since the 2014/15 season onwards.
}
\description{
Load shots data.
}
\examples{
shots_data <- load_shots()
}
