% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnblr-package.R
\docType{package}
\name{wnblr-package}
\alias{wnblr}
\alias{wnblr-package}
\title{wnblr: Data sets containing Women's National Basketball League game stats}
\description{
Contains four types of game stats from the Australian Women's National Basketball League (WNBL): team box scores, player-level box scores, play-by-play data, and shot location data. Currently, the package includes available data from games spanning the 2014/2015 to 2020 seasons, inclusive. Original data source: the WNBL (https://www.wnbl.com.au) and FIBA Livestats.
}
\author{
\strong{Maintainer}: Jacquie Tran \email{jac@jacquietran.com}

}
\keyword{internal}
