% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_scores.R
\docType{data}
\name{box_scores}
\alias{box_scores}
\title{Box score data from the Australian Women's National Basketball League (WNBL).}
\format{
A tibble with 1200 rows and 52 variables:
\describe{
\item{season}{numeric, season of competition. Note that WNBL seasons typically overlap year-end; for simplicity, the first year of competition is included in the \code{season} variable, e.g., "2014" represents the 2014 / 2015 season}
\item{page_id}{numeric, unique identifier for each game. View the game stats in your browser by navigating to https://www.fibalivestats.com/u/wnbl/<page_id>/}
\item{team_name}{character, team name}
\item{team_short_name}{character, team nickname}
\item{home_away_flag}{character, identifying whether the team represented on each row was the home or away team.}
\item{team_name_opp}{character, opposition team name}
\item{team_short_name_opp}{character, opposition team nickname}
\item{points_total}{numeric, points scored by the team listed in \code{team_name}}
\item{points_total_opp}{numeric, points scored by the opposition team listed in \code{team_name_opp}}
\item{points_margin}{numeric, \code{points_total} minus \code{points_total_opp}}
\item{team_result}{character, win ("W"), loss ("L"), or draw ("D")}
\item{minutes}{}
\item{field_goals_made}{}
\item{field_goals_attempted}{}
\item{three_pointers_made}{}
\item{three_pointers_attempted}{}
\item{three_pointers_percentage}{}
\item{two_pointers_made}{}
\item{two_pointers_attempted}{}
\item{two_pointers_percentage}{}
\item{free_throws_made}{}
\item{free_throws_attempted}{}
\item{free_throws_percentage}{}
\item{rebounds_defensive}{}
\item{rebounds_offensive}{}
\item{rebounds_total}{}
\item{assists}{}
\item{turnovers}{}
\item{steals}{}
\item{blocks}{}
\item{blocks_received}{}
\item{fouls_personal}{}
\item{fouls_on}{}
\item{fouls_total}{}
\item{points}{}
\item{points_from_turnovers}{}
\item{points_second_chance}{}
\item{points_fast_break}{}
\item{bench_points}{}
\item{points_in_the_paint}{}
\item{time_leading}{}
\item{biggest_lead}{}
\item{biggest_scoring_run}{}
\item{lead_changes}{}
\item{times_scores_level}{}
\item{fouls_team}{}
\item{rebounds_team}{}
\item{rebounds_team_defensive}{}
\item{rebounds_team_offensive}{}
\item{turnovers_team}{}
\item{efficiency_team}{}
}
}
\source{
Australian Women's National Basketball League website \url{https://www.wnbl.com.au}
}
\usage{
box_scores
}
\description{
Contains box scores data at the team-level (one row per competing team), from games played in the Australian Women's National Basketball League. Includes common basketball metrics such as points scored, rebounds, assists, steals, and so on. The data set currently contains publicly available data from games played in the 2014/2015 season to the 2020 season, inclusive.
}
\keyword{datasets}
